Patch #4891: Moved operator<< for Gaudi::Time in namespace Gaudi
The operator<<'s where in the global namespace, but they should be in
the same namespace as the class they operate on.
The presence of operator<< for StringKey and others in the Gaudi
namespace confused the compiler which was not able anymore to find the
ones in the global namespace.