Disabled automatic zipping of Python modules in CMake (ordering problem)
It's not possible to fix the order in which install steps are done, so
we cannot ensure that the zipping is executed _after_ the installation
of all Python files.

I added a custom target (python.zip) as a short-cut to call the command,
but it has to be run by hand.

See http://public.kitware.com/Bug/view.php?id=8438