prototype infrastructure to chain projects with subdir override
The regular "export" mechanism of CMake cannot be used because it
prevents the overriding of subdirs and it doesn't allow transitivity of
include directories.

The choice is to implement our own export mechanism that allow overrides
and exports all the special properties the framework is using
(transitive includes and libraries).

The main disadvantage is that the implicit transitive link has to be
reimplemented within our framework and thus we are bound to the
framework to create libraries, modules and executables.

TODO:
- the import of other projects (CMT 'use') has to be implemented and
  made recursive (the prototype uses hard-coded Gaudi to test)