Added the possibility to declare targets that generate headers.
If a subdir has targets that produce headers that may be used by other
subdirs, it must declare them in the property GENERATED_HEADERS_TARGETS
of the directory. The 'gaudi_add_' family of functions will collect
those targets from all the dependencies and declare the correct
dependency on the generated target.

It is also possible to use the variable required_local_genheader_targets
to enable the mechanism within a package.

From: Marco Clemencic <marco.clemencic@cern.ch>


git-svn-id: svn+ssh://svn.cern.ch/reps/gaudi/Gaudi/trunk@6985 53ecefc3-a64d-0410-9bdf-b8581fa3287e