added support for data packages
Data packages can be declared in the 'DATA' list of the arguments to
gaudi_project(), with or without version pattern, For example:

  gaudi_project(MyProject v1r0
                DATA SomeDataPkg
                     AnotherDataPkg VERSION v3r*)

The be found, the data package must contain an environment XML file
called <name>Environment.xml, where <name> is the name of the package
(with '/'s replaced by '_'s).

The search path for data packages is CMTPROJECTPATH + CMAKE_PREFIX_PATH,
with the suffixes declared in the (cached) variable GAUDI_DATA_SUFFIXES.
For example, if GAUDI_DATA_SUFFIXES contains "D1;D2", the packages are
searched for in:
  <prefix>/
  <prefix>/D1
  <prefix>/D2

By default GAUDI_DATA_SUFFIXES is set to "DBASE;PARAM;EXTRAPACKAGES"
(LHCb convention).