added code to force merge of files even if the build fails
In the old system, a partial/failed build produces partial merged files.
To reproduce the same behavior, I added a set of new targets
(Merged*_force) to trigger this partial generation by hand (for example
for a test) and a special install code to force the generation if the
output files are not found.