Added a thread safe detector data service
The idea is to replace the DataSvc with a version with thread safe
retrieve methods called TsDataSvc. The DetDataSvc now inherits from
this TS version of DataSvc. The RegistryEntry had to be made friend
of this new svc.
For the moment a recusrive lock is used given the "richness" of the
IDataManager interface.
Reentrant methods should be spot and recursive locks eliminated in
favour of a rapid spin lock. It is not clear to me yet if this can be
done without changing the interface.