Warn only if new number of slots is different from previous one
At the moment the number of slots can be set only via properties.
The setNumberOfStores method cannot be invoked after the GaudiStateMachine
is INITIALIZED.
The temporary usage therefore foresees the presence of the same parameter
for the scheduler, if any, and the whiteboard in the configuration.
A dynamic changing of stores at runtime will be addressed in the future.