Minimally upgrade ToolSvc to be able to properly clone private tools
A minimal change (< 20 lines) was committed to the tool svc in order
to clone private tools.
The mechanism allows to identify a tool with its parent (in addition
to name and type) when requested.
Any attempt to clone public tools would result in an unpredictable
behaviour.
The details about the modified files:
  o GaudiCoreSvc/src/ApplicationMgr/ToolSvc.cpp: actual changes in
the retrieve and create methods.
  o GaudiExamples/CMakeLists.txt,
    GaudiExamples/options/AlgToolsClone.py,
    GaudiExamples/src/AlgTools/MyAlgorithm.cpp,
    GaudiExamples/src/AlgTools/MyAlgorithm.h: changed to support a
test added in CTest
  o GaudiHive/src/AlgResourcePool.cpp: add initialisation for cloned
algorithms (forgot before, never noticed because we never cloned
anything besides crunchers before...)