Define GOD_NOALLOC variable to eliminate new/delete overriding in event model
The new and delete operators were overidden by default in the event data model
classes (LHCb/Event).
The allocator was replaced with the boost pool allocator, which works great
in the serial and in the non-Numa contexts but had huge implications when
the first allocation happened on the Dram associated to one processor and
a thread was using this pool. Probably quite some traffic was generated.
with the definition of this variable, the code for the overriding is
generated but never compiled (ifdef). As a result minibrunel scales
up to 5 instead of 4 on a dual processor, 4 cores numa machine.