Fixed bug https://sft.its.cern.ch/jira/browse/CFHEP-114 + code conventions
The bug affected the declaration of algorithms dependencies directly in
the scheduler. When the number of dependencies (different data objects)
was bigger than the number of algorithms, a read beyond the limit of a
vector happened. This has been fixed setting the proper vector size
and a test has been added to checkpoint this correct behaviour.

In addition some polishing took place to adapt to the coding conventions.