allow overrides of LCG_externals_*.txt
All the LCG_externals_*.txt in the CMAKE_PREFIX_PATH (+ CMTPROJECTPATH, with
optional suffix dir LCG_<version>) are taken into account to fill the
CMAKE_PREFIX_PATH for the externals.

The first LCG_Externals_*.txt in the path overrides the content of the
following ones, and the last one is used to determine the base location of
the externals.