Fixed GAUDI-1019: use relocatable path for gcc
I modified the relocation for new style LCG releases
(lcg/releases/LCG_<version>) to use a new variable (LCG_releases_base)
pointing to the base directory of all LCG releases (lcg/releases).

In conjunction with the fix to LBCORE-783 (or by using the copy of gcc
in lcg/releases), the path to gcc becomes relocatable.