revert declaring constructors as = delete as this prevents the generation of the python dictionaries.
Declare them once more as private, and do not implement instead.

Note: there is something funny with the dependencies: make the
'working' version of the code, then revert this patch, and build again,
and the python dictionaries are not redone. Then do 'make purge'; 'make install'
and it fails to compile...