fixed double delete of tools
With 5e26eee, the data member `m_tools` was added to `AlgTool`, from the
equivalent in `GaudiCommon`.
`m_tools` was used in `GaudiCommon` to record the tools retrieved via
the 'tool' method, but also to keep track of the tools managed with
`ToolHandle`s, so, during finalize, `GaudiCommon` was releasing a tool
that might have been released already by the handle.

I modified `GaudiCommon` so that `m_tools` is used for all tools, while
a new data member (`m_managedTools`) for the toos `GaudiCommon` must
release.