Added terminatePool to ThreadPoolSvc
Added the terminatePool functionality to the ThreadPoolSvc to be able to
invoke the IThreadInitTool::terminateThread method analogously to how
the initThread is called concurrently on each thread.

Added terminatePool method to the IThreadPoolSvc.

Refactored the ThreadPoolSvc so that the code to launch the
tool-wrapping tasks is in a dedicated method, launchTasks, which can be
called from both initPool and terminatePool. The boost::barrier is now a
function-local instance rather than a class member pointer.