Moved terminatePool call to end of activate method
Realized after testing that the IThreadPoolSvc::terminatePool call needs
to be executed in the scheduler thread, just like the initPool call.
Otherwise, the ThreadInitTools won't be terminated on all threads. This
may be due to a subtle feature of TBB which would be good to understand,
but this way the implementation probably makes more sense anyway.

Added an additional error check in ForwardSchedulerSvc::finalize to
catch any problems that may have occurred during the call to
terminatePool.