Initializes handles in the sysInitialize method of AlgTool and Algorithm
Handles in algorithms and algTools were initialized on demand inside the get method. This commit moves the intialization to the initialize phase.
This is part of an effort to make the get method const.

On my way, I've also factorized out some code concerning the handles between AlgTool and Algorithm into a DataHandleHolderBase class