use boost::string_ref to hold Property name and documentation
The use of boost::string_ref (to be replaced by std::string_view) reduce
significantly the space required by Property, avoiding pointless
duplications in memory (see GAUDI-1229).

With the `profile_Property` executable I measured a reduction of a
factor 2 in size (272 -> 128 bytes) with a small increase in time
(0.2 -> 0.3 us) for each created Property (always the same name and doc).