Enhance SplittingTransformer
In order to allow the SplittingTransformer to make a distinction between
'no output' and 'empty output', recognize whether the output is a vector
of optionals. If so, then only put an output element into the event
store if the optional 'contains' something (which may be an empty
container!), i.e. preserve the 'optionalness'.

This is done so that one can implement the following pattern: in the
HLT decoders, one loops over the rawbank, and looks for a given number
of containers. These containers could be present and non-empty, present
and empty, or non-present. The first two cases are treated the same, and
are the regular case. But in case of absence in the raw data this should
be reflected in absence in the event store.