hide missing override warning coming from a CLHEP macro
Until we pick up a more recent CLHEP (see CLHEP-136) we need to hide the
warning.

Hiding the warning here is special because we need different pragma lines
for GCC or Clang, and that of Clang should also be used when the header is
processed by Cling (and Cling pretends to be GCC).