add sysExecute(EventContext) and extract errorCount from Algorithm base class
We need to extract the m_errorCount variable from the Algorithm base class,
as it's not important on a per-instance basis, but rather on a job level. This
is moved to the AlgExecStateSvc.

This merge request also adds an explicit Algorithm::sysExecute(const EventContext&),
which should become the standard entry point to Algorithm::execute. This sets the local
m_event_context variable, so derived classes can get quick access to the current context
instead of going through the slower thead local Gaudi::Hive::currentContext.