Made StatEntity more thread safe
Note that it says more. This is because there is a trade-off here between full thread safety and performance.
Namely, what I've done is adding a lock when writing to counters. So counters will have now the proper value in all cases.
But it does not mean that the whole class is thread safe. In particuler querying while updating may not give consistent answers as the update itself is not locked and several internal counters are potentially queried.
As an example, querying the mean involves dividing the sum counter by the nbcalls counter, and one may have been updated and not yet the other.

However this is felt to be the best compromise to not introduce heavy locks everywhere in the code