Ignore all Data Dependency calculations if Algorithm is a Sequence
If an Algorithm is a Sequence, the scheduler pays no attention to the
data dependency rules, and just does control flow. So there's no need
to do any data dependency calculations, or filter the dependency loops,
for which, by definition, there will be.

The ugly bit is that we have to check for the existence of the Property
"IgnoreFilterPassed", as there's no good way of determining if an
Alg is a Sequence.