Sort scheduler debugging dumps in a reproducible order.
In several places, AvalancheSchedulerSvc dumps out sets of DataObjID objects
by iterating over an unordered_set of them.  However, the ordering of this
iteration is undefined and can, and does, vary between different compilers
and different versions of the same compiler.  This means that the ordering
of the output can vary depending on the compiler, which is inconvenient
for regression tests.

This change adds a sorting step so that the DatObjID sets are printed
in a well-defined order.  This only affects the dumps, so should
not be a performance issue.