Reallowed global postprocessing in ScalarTransformer
This had been introduced and removed in a6d51f57 as it was creating recursion issues in case the In and Out parameters of the template are the same.
This was basically do to a reuse of operator() for 2 things : the processing and the postprocessing.
Now postprocessing is using the named function postprocess and everything works fine