Extend AlgExecState with an "executing" state
The AlgExecState, accessible via the IAlgExecStateSvc, now records if an
algorithm is currently being executed. This required some interface
changes to the AlgExecState, i.e. replacing the boolean [set/is]Executed()
with a enum-based [set]execState(). Clients have been updated
accordingly. This also opens the door for further algorithm states if
necessary later.