Flesh out the DataHandle implementation further
Upstream DataHandle now has all the information it needs to inform
future Property implementation of access mode and data store specifics.

There is a sketch of how DataHandles will be made compatible with
multiple data stores at a later point in time.

DataHandles now support both insertion of pointer-to-DataObject and
insertion of arbitrary values. There is a sketch of how automatic
casting of containers to equivalent ranges will eventually work.