Fixed race condition in AvalancheScheduler around updates after execution
In methods promoteTo(Async)Executed, 2 types of "updates" are taking place :
  - a push to the actionQueue for the updates of other algorithms
  - the update of the current algorithm status
The push itself will ultimately trigger the check for the end of the event, that will check whether all states of all algos of the event are ACCEPTED.
In such a case, it will trigger (via a push to m_finishedEvents queue) the move to the next event or the finalization of the job (last event).
Now suppose that the 2 updates are done in the order I gave them and that the second one is delayed (thread unscheduled by OS).
The push will trigger the check for end event. That check will concledu that the event is not over as one algo (current one) is still in SCHEDULED state !
Thus next event will never come and job will never finalize.
So the order has to be update of the current algo first.

Note that this was found the hard way, on stuck processes, so this really happenned.