Simplifies (aka optimized) usage of ttb tasks in AvalancheScheduler
Each algorithm was so far given to tbb in the form of a pair of tbb tasks, linked together.
The first was running the algo while the second dealt with pushing the updae callback to the action queue.
Providing that the run of the Algorithm was already wrapped into a AlgoExecutionTask, there was no reason to not put the callback there and drop the second task.