Added protection on event slot creation in HiveWhiteBoard
So far the code could try to create a new event even if no slots were available.
This was due to the fact that there was not count of free slots.
The consequences where an exception out of range further down the road as the return code of the event creation was on top not checked properly.
This adds a free slot counter in the white board