Make Gaudi::Property<T> work even if T is only copy-constructible
Right now, the Gaudi::Property<T> code assumes that the type T is
default-constructible. But it only uses this assumption when a value of
type T is parsed from a string. With a couple of modifications to the
string parsing function, Gaudi::Property<T> can work with
copy-constructible types as well.