More standard compliant type-punning
Type punning was implemented using a `union`, which relies on
writing to one member of the union, and then reading from another.
Strictly speaking this is 'undefined behaviour' in C++, as only
the 'active' member of a union should be used. However, this
works as a gcc extension, as it is valid C11 (note: no ++!).
This patch changes this to use `std::memcpy`, which casts (implicitly)
to `unsigned char*`, and casting any pointer to/from `unsigned char*`
is well-defined in C++.

Note that in practice, gcc and clang generate _identical_ code -
see eg. https://godbolt.org/g/3kwRbo - so from that point
of view this patch 'does nothing' but it is more standards compliant ;-)