Optimized allocation of TES slot in HiveWhiteBoard
This was doing a linear search on the slots, taking a lock on each of them to check their availability.
Not only the algorithm was O(n) while it could be O(1) but the locking was not scaling when many threads were at work.
Putting a fifo queue with its own small lock solves the issue.