Added a new event loop manager for HLT like usage, that is very simple and very fast scheduling
This new event loop manager has an integrated scheduler that is very simple : it sequentializes the list of algorithms to run according to Data Dependencies and schedules full events in tbb, so that one task runs serially all algorithms one after the other.
The scheduling of events is pure fifo : schedule as many events as slots in the white board initially and triggers the next one when oen finishes.
In comparison with the Avalanche scheduler, many features are not present : no resources, no dynamic rearrangement of the sequence, no IO bound algos in particular.
On the other hand, the scheduling is much faster and allows to sustain high rates of events.