Add documentation README to THistSvc folder
Add legacy methods for regGraph and regTree

Migrating the Athena in these cases would be possible, but not very
clean. This is why I decided to favor a consistent interface (with
similar methods for Hists, Trees and Graphs), opposed to only providing
unique_ptr methods. It would be good if future THistSvc clients would
pick up the new methods and old methods would be migrated over time.