Fix bug in deReg and add test
The lambda in deReg( TObject* ) tried to access vhid->at(0) after a
check for vhid->size() == 0. This resulted in a certain crash.

I've reorganized the function to delete the vector of THistIDs when
vhid->size() is 1 (not 0).
There is also a new test case that registers and deregisters a single
histogram.

Apply formatting