Remove inputDataObjs and outputDataObjs from DataHandleHolder
The original data handle holder was solely focused on event data, and
did not acknowledge the existence of conditions. Introducing conditions
in Gaudi therefore requires clarifying the distinction.

Since it was agreed at a Gaudi workshop that the distinction should not
be made at the DataObjID level, it must be made in functions that emit
or accept DataObjIDs.