ConcurrencyFlags: move setting from AvalancheScheduler to HiveWhiteBoard
Move the setting of the number of stores in ConcurrencyFlags from the
scheduler to the whiteboard. This is more logical and also makes the
correct number of slots available before the scheduler is initialized
(e.g. in the ATLAS HLT we initialize the scheduler only after the start
transition to avoid creating thread pools before forking).