THistSvc: fix memory corruption in deReg(string)
Deregistering a histogram via its name causes a memory corruption (often
resulting in a crash). The problem was that deReg(string) kept checking
the histogram list size while the list is actually being deleted in the
last call to deReg(TObject*).

Also improved the unit test coverage for this case.