Add Property constructors which accept an updateHandler
The updatehandler can either be anything which satisfies
the `void(PropertyBase&)` signature, or a member function
with the signature `void()` or `void(PropertyBase&)`.
In addition, an additonal argument can be passed which specifies
whether or not the provided handel should be invoked immediately.
Default is to _not_ invoke it immediately.