Fix MessageSvc::reinitialize
The addition of the call to `MessageSvc::reinitialize` in gaudi/Gaudi!866 exposed a bug
in that method leading to
```
MessageSvc          ERROR Gaudi::StateMachine::ChangeState   Invalid
transition 'START' from state 'CONFIGURED'   StatusCode=FAILURE
```
when `ApplicationMgr.reinitialize()` is called.

Setting the target state explicitly and add a unit test to check the
basic state transition for `ApplicationMgr`.