Re-designed how GaudiPluginService would use std::filesystem.
At the moment only GCC 8+ and Clang 7+ implement std::filesystem from the
C++17 standard. I decided that the code should only use std::filesystem when
one of these compilers is used for the build, and for all other cases fall
back on boost::filesystem. This allows for a successful build on some
non-standard platforms, macOS most of all. ;-)