Specify JOB_POOL for genreflex custom commands
A variable hook `GENREFLEX_JOB_POOL` is added,
that allows limiting the parallelization of genreflex commands
independently of the default parallelism when using ninja.
See the ninja [manual](https://ninja-build.org/manual.html#ref_pool)
for more details on job pools. This is especially needed
when distributing the compilation as the number of jobs is very high
and trying too many genreflex calls locally can use up the memory.

The new variable can be set from the toolchain or the pre-loaded initial
cache.