Made the usage of std::filesystem configurable.
In certain situations it is necessary to overrule the decision of the
Gaudi code on whether std::filesystem is safe to use or not. For
such cases there is now a cache variable (GAUDI_USE_STD_FILESYSTEM)
in place, that one can use from the outside to control this
behaviour.