Improvements in Accumulators to increase genericity
The goal was to be able to define Accumulators having several Arithmetic
types in view of an histograming counter. This had a few consequences :
  - Buffer and Bufferable counters needed a variadic template for the
    Arithmetic types
  - thus this had to be the last parameters in the template parameters
  - and thus Accumulators had also to change their template parameter
    order and put Arithmetic type(s) as the last one(s)

On the way, the inheritance diagram was slightly simplified between
Counters and Accumulators, moving the inheritance on Accumulators to the
BufferableCounter class from the user facing Counter class. This allowed
to drop a static_cast.

merge accumulators