Implemented printing of counters through the Monitoring::Hub
Note that MonitoringHub is now Gaudi::Monitoring::Hub.

Counters are now serialized in json and sent to a MessageSvcSink for printing.
This should be kind of backward compatible with current printing except that the Counter order may change.

The CounterHolder wrapper, now useless, has been dropped and counters register directly to the Monitoring::Hub
Note that his had the consequence that the Counters method of the python interface had to be dropped.
It does not seem to be used, but in case it would be needed, it can be implemented if we add a dedicated method to list counters of a given component to the Monitoringi::Hub