Add support to DataWriteHandle for writing 'views' into the TES
Allow a second template argument on `DataObjectWriteHandle` eg.
DataObjectWriteHandle<GetType,PutType> (note: by default, in case
of a single template argument, PutType=GetType, and the original
behaviour is kept).  In this case, `DataObjectWriteHandle::put`
will accept an rvalue reference to `PutType`, which it will create
a dedicated 'wrapper', which contains both the moved instance of
`PutType` as well as a `GetType` constructed from the moved instance,
into the event store. This wrapper can be retrieved and unwrapped
by a `DataObjectReadHandle<GetType>`.