Fix GaudiPython classes (follows up !1116)
Fix the workaround introduced in 4f05f0367897e3d5b88ebd64b95bd2dc470261a4
(for the lack of mixed C++/python inheritance) as part of !1116.
There, the `iAlgorithm` base class was made into a member but the
semantics of attribute lookup were not kept backward compatible
(and e.g. the `.Histos` member of `iAlgorithm` was not accessible).