IncidentSvc: clear incident map
The map used for the storage of the asynchronous incidents was never
cleared during the job and each new `EventContext` added a new entry
into the map that only got cleared during `finalize()`.

Change the storage to be slot-specific and clear the list of incidents
when the event changes for a given slot.