Do not clearStore too early in EventLoopMgr
Before this change, there used to be an inconsistent behavior when using
GaudiPython concerning the state of the TES after a call to appMgr.run.
In case we had reached the end of input, it was empty as clearStore was
called. But in case we did not, it was populated with the last event.

Providing the latter case allows to run more algorithms by hand or dump
content from the TES (used for example by the LHCb alignment code), this
change removes the clearStore call in case we reach the end of input.

Note that it does not change the normal behavior as it is anyway called
in the next few lines in case a new event comes, through the set Root
call