Use FIXTURES_REQUIRED for QMTest test dependencies
For a nice introduction to `FIXTURES_*` in CTest see
[here](https://crascit.com/2016/10/18/test-fixtures-with-cmake-ctest/).

`FIXTURES_REQUIRED` is set in addition to setting `DEPENDS` for ever test that
defines `prerequisites`. Using `FIXTURES_SETUP` a fixture is defined for every
test that is a dependency of another, where the fixture name is identical to
the test name.

The main advantage of using fixtures is usability: `ctest -R <regex>`
will also run the dependencies, unless explicitly ignored (e.g. with `-FA .*`).
Moreover, when a dependency (fixture) fails, the dependent test is skipped
("Not Run").