Introduced DeprecatedDynamicDataObjectHandle
This transitional class will help moving all code to Handles even if
some usage requires 'dynamic' behavior, i.e. to declare handles at run
time only and not registering them with their owner.

Again should no be used for long term code, only for helping in the
transition. It is bound to be marked Deprecated and then dropped as
the name explicitely says.

Note that it needed a new constructor in the DataObjectHandleBase but
that one is directly marked deprecated and used only in the new class
where the deprecation warning is ignored