Added removeEntity interface to MonitoringHub
This allows to remove entities from the hub (and thus from the sinks in
the back) and thus allows to use entities in temporary objects, and not
only in Tool/algorithms.

The counter part is that BufferableCounters are no more movable,
although I doubt anyone really want to move a counter. It only has the
consequence that you can hardly have vectors of counters, but arrays or
deques are still fine