Add a MultiMergingTransformerFilter.
Based on the MultiMergingTransformer implementation, but allows for
algorithms to return a boolean which sets the 'filter passed' property
for the control flow.

A new test checks an example instantiation along one of
MultiMergingTransformer, which was not previously tested. To make this
test more obvious the IntDataProducer example now has a configurable
value property.