"Fixed" scan_dict_deps.py for not including directories
This was discovered thanks to a name collision between Eigen and DD4hep,
Eigen having an header file named "Core" (with no extension indeed) and
DD4hep a directory with same name. DD4hep happened to come first in the
dependency list and thus the directort was used...

Note that this reveal a deeper problem in the scanning made by this tool
: no scoping and thus no cleaning of the list of directories when going
to next file. This is the actual root of the problem that lead to DD4hep
being first in the dependency list