Make Gaudi::Functional constructors more paranoid about # of inputs/outputs
Currenly, the constructors of the Gaudi::Functional baseclasses use
`std::array<KeyValue,N>` as arguments where N is the number of inputs
or outputs specified. However, this implies that one can specify a list
of less than N KeyValues, as `std::array` will happily
default-initialize any missing items. This will then result in an error
in genConf, as this implies a property with a default-initialized string
(i.e. an empty string) as key.

This can be fixed by replacing the `std::array<KeyValue,N>` with a tuple
with N `KeyValues, eg. for N=2 an `st::tuple<KeyValue,KeyValue>`.