Changes to the `Configurable` method `merge`
The changes applied to the `merge` method
were made as a result of code profiling
and result in a roughly 20-30s improvement
in some ATLAS trigger jobs.

An new test was added in
`test_configurables_merge.py`
as a result of changes made to `merge` in order
to increase test coverage.