Adds CheckedNamedToolsConfigured functionality
This checks whether an explicitly named tool has had an property configured.
This is an attempt to detect a category of error where a tool is configured, but not added as a public tool.
In which case an unconfigured default tool is created, BUT with the expected non-default name, causing confusion.